Add Conditional Logic to Decide When to Add Presentment Prices Header
I have added a considerably bloated conditional check to only add the
header for the products, product_listing and product_variants APIs.
I have updated the tests to validated that the header is not added
for other types of API calls and that the expected header is present
for the GET requests pulling product related data.