feat: add an option for automatic retries when hitting rate limits
This adds opt-in retry logic to calls made to REST or GraphQL endpoints on Shopify. Many users of the Shopify API run into the restrictive rate limits, and there's one simple way you can work around them, which is to retry the request after the time has elapsed. This isn't the only strategy, but it's a really common one because it doesn't require any cross-process coordination or state. Instead of trying to guess before hand if a request is going to work or not (which is hard to do if you have many different processes on many different servers all making requests at the same time), this will just always try the request, and keep retrying for a while until it succeeds. This is simple and robust in my experience, and so I think is a great candidate for inclusion in this library.

This approach uses Got's built in `retry` configuration option, so `shopify-api-node` doesn't have to implement any of the custom backoff logic or actual retry code, and instead just has to compute the time to wait before retrying which is kind of nice.

I didn't turn on this behaviour by default so as to not make a breaking change.